Hide TcpTransport being detected as a standalone transport.
