+++ docs/index.html
@@ -8025,7 +8025,7 @@ function sourceButton(element) {
             <span class="type-default">0</span>
 
                 </div>
-                <div class="signature-description desc"><p>Number of retries to make when a put failure occurs.</p>
+                <div class="signature-description desc"><p>Number of retries to make when a write failure occurs.</p>
 </div>
             </div>
             <div class="signature-box-container">
@@ -8139,7 +8139,7 @@ function sourceButton(element) {
       "minimum": 0,
       "maximum": 10,
       "default": 0,
-      "description": "Number of retries to make when a put failure occurs."
+      "description": "Number of retries to make when a write failure occurs."
     },
     "retry_delay": {
       "type": "integer",
@@ -10487,7 +10487,7 @@ function sourceButton(element) {
           "minimum": 0,
           "maximum": 10,
           "default": 0,
-          "description": "Number of retries to make when a put failure occurs."
+          "description": "Number of retries to make when a write failure occurs."
         },
         "retry_delay": {
           "type": "integer",
